Overriding styles of components is done via overriding the theme per component
Styles do not map one-to-one with elements within the component and would require knowledge of internal workings of component to apply custom ones. The theme represents the configurable parts of the component's visuals, so it makes sense that overriding the visuals per component would mean overriding a theme prop on each component.