D-12711 Create scenario for issue via a Popover story
Scenario exists when the popover is configured to **not** close when the anchor exits the screen. In these cases, the exhibited behavior is that the popover will remain on the screen; however, it will remain fixed to the top of the screen.

Desired behavior is for it to remain next to its anchor and for both to scroll off the screen.